# gmc_api_gateway/docs/swagger.yaml
basePath: /gmcapi/v2
definitions:
...
  model.Conditions:
    properties:
      lastProbeTime:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  model.DEPLOYMENT_DETAIL:
    properties:
      NodeSelector: {}
      annotations: {}
      cluster:
        type: string
      containers: {}
      createAt:
        type: string
      events:
        items:
          $ref: '#/definitions/model.EVENT'
        type: array
      labels: {}
      name:
        type: string
      project:
        type: string
      ready:
        type: string
      replica:
        $ref: '#/definitions/model.REPLICA'
      strategy: {}
      updateAt:
        type: string
      user:
        type: string
      workspace:
        type: string
    type: object
...
  model.EVENT:
    properties:
      cluster:
        type: string
      eventTime:
        type: string
      kind:
        type: string
      message:
        type: string
      name:
        type: string
      namespace:
        type: string
      reason:
        type: string
      type:
        type: string
    type: object
  model.Error:
    properties:
      data:
        description: Email    string `json:"email"`
        type: string
      status:
        type: string
    type: object
...
  model.JOB:
    properties:
      cluster:
        type: string
      completions:
        type: string
      created_at:
        type: string
      duration: {}
      name:
        type: string
      project:
        type: string
      user:
        type: string
      workspace:
        type: string
    type: object
  model.JOB_DETAIL:
    properties:
      annotations: {}
      backoffLimit:
        type: integer
      cluster:
        type: string
      completionTime:
        type: string
      completions:
        type: string
      conditions:
        description: OwnerReference []OwnerReference `json:"ownerReferences,omitempty"`
        items:
          $ref: '#/definitions/model.Conditions'
        type: array
      containers: {}
      created_at:
        type: string
      duration: {}
      events:
        items:
          $ref: '#/definitions/model.EVENT'
        type: array
      label: {}
      name:
        type: string
      parallelism:
        type: integer
      project:
        type: string
      startTime:
        type: string
      status:
        type: integer
      user:
        type: string
      volume:
        type: integer
      workspace:
        type: string
    type: object
...
  model.REPLICA:
    properties:
      availableReplicas:
        type: integer
      readyReplicas:
        type: integer
      replicas:
        type: integer
      unavailableReplicas:
        type: integer
      updatedReplicas:
        type: integer
    type: object
...
  model.WORKLOAD:
    properties:
      NodeSelector: {}
      cluster:
        type: string
      createAt:
        type: string
      name:
        type: string
      project:
        type: string
      ready:
        type: string
      user:
        type: string
      workspace:
        type: string
    type: object
...
host: 101.79.1.138:8012
info:
  contact:
    name: GM-Center
    url: https://gedge-platform.github.io/gm-center/
  description: This is a Gedge GM-Center Swagger API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gedge GM-Center Swagger API
  version: "2.0"
paths:
...
  /deployments:
    get:
      consumes:
      - application/json
      description: get Deployment List
      parameters:
      - description: name of the Workspace
        in: query
        name: workspace
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        type: string
      - description: name of the Project
        in: query
        name: project
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WORKLOAD'
      security:
      - Bearer: []
      summary: Show List Deployment
      tags:
      - Kubernetes
    post:
      consumes:
      - application/json
      description: Create Deployment
      parameters:
      - description: Deployment Info Body
        in: body
        name: json
        required: true
        schema:
          type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WORKLOAD'
      security:
      - Bearer: []
      summary: Create Deployment
      tags:
      - Kubernetes
  /deployments/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Deployment
      parameters:
      - description: name of the Deployment
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - Bearer: []
      summary: Delete Deployment
      tags:
      - Kubernetes
    get:
      consumes:
      - application/json
      description: get Deployment Details
      parameters:
      - description: name of the Deployment
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DEPLOYMENT_DETAIL'
      security:
      - Bearer: []
      summary: Show detail Deployment
      tags:
      - Kubernetes
  /jobs:
    get:
      consumes:
      - application/json
      description: get Job List
      parameters:
      - description: name of the Workspace
        in: query
        name: workspace
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        type: string
      - description: name of the Project
        in: query
        name: project
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JOB'
      security:
      - Bearer: []
      summary: Show List Job
      tags:
      - Kubernetes
    post:
      consumes:
      - application/json
      description: Create Job
      parameters:
      - description: Job Info Body
        in: body
        name: json
        required: true
        schema:
          type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JOB'
      security:
      - Bearer: []
      summary: Create Job
      tags:
      - Kubernetes
  /jobs/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Job
      parameters:
      - description: name of the Job
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - Bearer: []
      summary: Delete Job
      tags:
      - Kubernetes
    get:
      consumes:
      - application/json
      description: get Job Details
      parameters:
      - description: name of the Job
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JOB_DETAIL'
      security:
      - Bearer: []
      summary: Show detail Job
      tags:
      - Kubernetes
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: '"Type \"Bearer \" and then your API Token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"